Group3
In recent years, medical image classification has emerged as a crucial field in healthcare, leveraging the power of machine learning and artificial intelligence to aid in disease diagnosis and treatment. With the advent of advanced imaging technologies, such as MRI, CT scans, and X-rays, the amount of medical image data has increased exponentially.